Swap to docusaurus serve + node alpine for image
This results in significantly larger images as we are no longer just packaging up the build artifact files, but may fix issues with nginx performing 301s due to SSL termination and breaking some desired indexing for docs.

If it doesn't, will simply revert to the old image and investigate other solutions.